+++ evalml/tests/component_tests/test_components.py
@@ -26,6 +26,7 @@ from evalml.pipelines.components import (
     ComponentBase,
     DateTimeFeaturizer,
     DelayedFeatureTransformer,
+    DFSTransformer,
     DropColumns,
     DropNullColumns,
     ElasticNetClassifier,
@@ -33,7 +34,6 @@ from evalml.pipelines.components import (
     Estimator,
     ExtraTreesClassifier,
     ExtraTreesRegressor,
-    FeatureTools,
     Imputer,
     LightGBMClassifier,
     LinearRegressor,
@@ -147,7 +147,7 @@ def test_describe_component():
     text_featurizer = TextFeaturizer()
     lsa = LSA()
     pca = PCA()
-    ft = FeatureTools()
+    ft = DFSTransformer()
     assert enc.describe(return_dict=True) == {'name': 'One Hot Encoder', 'parameters': {'top_n': 10,
                                                                                         'features_to_encode': None,
                                                                                         'categories': None,
@@ -169,7 +169,7 @@ def test_describe_component():
     assert text_featurizer.describe(return_dict=True) == {'name': 'Text Featurization Component', 'parameters': {'text_columns': None}}
     assert lsa.describe(return_dict=True) == {'name': 'LSA Transformer', 'parameters': {'text_columns': None}}
     assert pca.describe(return_dict=True) == {'name': 'PCA Transformer', 'parameters': {'n_components': None, 'variance': 0.95}}
-    assert ft.describe(return_dict=True) == {'name': 'Featuretools', 'parameters': {"max_depth": 2}}
+    assert ft.describe(return_dict=True) == {'name': 'DFS Transformer', 'parameters': {"index": "index"}}
 
     # testing estimators
     base_classifier = BaselineClassifier()
@@ -534,7 +534,7 @@ def test_transformer_transform_output_type(X_y_binary):
                 assert transform_output.shape[0] == X.shape[0]
                 assert transform_output.shape[1] <= X.shape[1]
                 assert isinstance(transform_output.columns, pd.Index)
-            elif isinstance(component, FeatureTools):
+            elif isinstance(component, DFSTransformer):
                 assert transform_output.shape[0] == X.shape[0]
                 assert transform_output.shape[1] >= X.shape[1]
                 assert isinstance(transform_output.columns, pd.Index)
@@ -556,7 +556,7 @@ def test_transformer_transform_output_type(X_y_binary):
                 assert transform_output.shape[0] == X.shape[0]
                 assert transform_output.shape[1] <= X.shape[1]
                 assert isinstance(transform_output.columns, pd.Index)
-            elif isinstance(component, FeatureTools):
+            elif isinstance(component, DFSTransformer):
                 assert transform_output.shape[0] == X.shape[0]
                 assert transform_output.shape[1] >= X.shape[1]
                 assert isinstance(transform_output.columns, pd.Index)